# Refactor LinearCombination