Follow up to https://github.com/invenia/FeatureTransforms.jl/pull/45 

This PR aligns the convention for `dims` used by `LinearCombination` with that of the rest of the package.

It also decouples the tests to make them independent with some minor refactoring to the code (renaming `x` -> `A`) to make it more consistent with other files.

There is one other change implemented in https://github.com/invenia/FeatureTransforms.jl/pull/47/commits/522043004a15ebcf743ae33737322df226b91779, which might be controversial and/or out of scope but it came up in the refactoring so I'll put it here for comment.

The basic idea is that a reduction operation should reduce the dimension of the input by 1. 
That is, an N-dimensional array input should be reduced to an (N-1)-dimensional output.

This really only gets interesting when the input is a vector and N=1. Intuitively, one expects the result of this to be a scalar. But Julia actually returns a 0-dimensional array, which as far as I understand are like magic scalars? As in, you can pass them to another function by reference but they also support array-arithmetic. See:
* https://github.com/invenia/FeatureTransforms.jl/pull/47/commits/522043004a15ebcf743ae33737322df226b91779
* https://discourse.julialang.org/t/what-is-the-reason-for-a-zero-dimensional-array/7175
* https://discourse.julialang.org/t/ref-vs-zero-dimensional-arrays/24434

(I think this also makes this function type-stable since it always returns an Array?)

In any case, `main` currently returns a 1-d vector, which is definitely inconsistent given the reasoning above if you ask me.
The question is: do we trust Julia to return a 0-dim array, or do we want to override it and return a scalar explicitly?

## background

Note that we want to emulate `sum(A; dims=d)` behaviour, that is, compute the linear combination and reduce over dimension `d`.

```julia
julia> M = reshape(1:9, 3, 3)
3×3 reshape(::UnitRange{Int64}, 3, 3) with eltype Int64:
 1  4  7
 2  5  8
 3  6  9

julia> sum(M; dims=1)
1×3 Array{Int64,2}:
 6  15  24

julia> sum(M; dims=2)
3×1 Array{Int64,2}:
 12
 15
 18

```

`main` currently gives the opposite result:
```julia
julia> lc = LinearCombination([1, 1, 1])
LinearCombination(Real[1, 1, 1])

julia> FeatureTransforms.apply(M, lc; dims=1)
3-element Array{Int64,1}:
 12
 15
 18

julia> FeatureTransforms.apply(M, lc; dims=2)
3-element Array{Int64,1}:
  6
 15
 24
```

this branch gives correct result:
```julia
julia> FeatureTransforms.apply(M, lc; dims=1)
3-element Array{Int64,1}:
  6
 15
 24

julia> FeatureTransforms.apply(M, lc; dims=2)
3-element Array{Int64,1}:
 12
 15
 18
```

The changes also makes this much more performant. 
See [previous benchmarks](https://github.com/invenia/FeatureTransforms.jl/pull/25#issuecomment-784184604) for setup:
```julia
# apply: this branch
0.446041 seconds (1.70 M allocations: 89.214 MiB, 6.05% gc time)
0.000210 seconds (912 allocations: 206.688 KiB)

# apply: main
0.339547 seconds (1.27 M allocations: 65.669 MiB, 8.31% gc time)
0.000913 seconds (30.61 k allocations: 812.016 KiB
```