@glennmoy 2021-03-10
> Balance is restored - looks good. I don't mind the 0-dimensional array. I just think the default dimension should be flipped.

> Should this also have a version bump?

yes, I'm tempted to say patch as this is fixing an inherent inconsistency we should have solved, that includes the 0-dim change.